Move services into service package
