+++ app/src/main/java/com/aconno/sensorics/service/AlarmService.kt
@@ -1,4 +1,4 @@
-package com.aconno.sensorics
+package com.aconno.sensorics.service
 
 import android.app.Notification
 import android.content.BroadcastReceiver
@@ -9,6 +9,7 @@ import android.os.IBinder
 import android.telephony.TelephonyManager
 import androidx.annotation.Nullable
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
+import com.aconno.sensorics.CpuWakeLock
 import com.aconno.sensorics.domain.AlarmServiceController.Companion.ACTION_ALARM_SNOOZE
 import com.aconno.sensorics.domain.AudioAlarm
 import com.aconno.sensorics.domain.telephony.CallStateListener
