+++ app/src/main/java/com/aconno/sensorics/service/AlarmServiceControllerImpl.kt
@@ -1,4 +1,4 @@
-package com.aconno.sensorics
+package com.aconno.sensorics.service
 
 import android.content.Context
 import android.content.Intent
