+++ app/src/main/java/com/aconno/sensorics/ui/MainActivity.kt
@@ -19,6 +19,7 @@ import com.aconno.sensorics.R
 import com.aconno.sensorics.domain.model.Device
 import com.aconno.sensorics.domain.scanning.BluetoothState
 import com.aconno.sensorics.domain.scanning.ScanEvent
+import com.aconno.sensorics.service.BluetoothScanningService
 import com.aconno.sensorics.ui.dashboard.DashboardFragment
 import com.aconno.sensorics.ui.device_main.DeviceMainFragment
 import com.aconno.sensorics.ui.devices.SavedDevicesFragment
