+++ app/src/main/java/com/aconno/sensorics/ui/device_main/DeviceMainFragment.kt
@@ -23,6 +23,7 @@ import com.aconno.sensorics.domain.interactor.filter.FilterByMacUseCase
 import com.aconno.sensorics.domain.model.Device
 import com.aconno.sensorics.domain.model.GattCallbackPayload
 import com.aconno.sensorics.domain.model.Reading
+import com.aconno.sensorics.service.BluetoothConnectService
 import com.aconno.sensorics.ui.ActionListActivity
 import com.aconno.sensorics.ui.MainActivity
 import com.aconno.sensorics.ui.beacon_settings.BeaconSettingsActivity
