+++ app/src/main/java/com/aconno/sensorics/viewmodel/MqttVirtualScanningViewModel.kt
@@ -3,7 +3,7 @@ package com.aconno.sensorics.viewmodel
 import android.content.Intent
 import androidx.lifecycle.AndroidViewModel
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
-import com.aconno.sensorics.MqttVirtualScanningService
+import com.aconno.sensorics.service.MqttVirtualScanningService
 import com.aconno.sensorics.SensoricsApplication
 import io.reactivex.disposables.CompositeDisposable
 import timber.log.Timber
